requirements/dev.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-solo==2.0.0
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg2-binary==2.9.3
pydantic==1.9.0
sqlparse==0.4.2
typing-extensions==4.1.1
